requirements.txt
# Environment
python-dotenv>=1.0.0

# Telegram
pyTelegramBotAPI>=4.31.0

# MongoDB
pymongo>=4.12.0

# LLM — calls LiteLLM proxy via OpenAI-compatible API
openai>=1.65.0

# Pydantic
pydantic>=2.10.0

# Web Search
tavily-python>=0.7.0

# File Processing
pypdf>=6.7.0
python-docx>=1.2.0
Pillow>=12.1.0          # requires Python >=3.10
pytesseract>=0.3.10

# Scheduling (reminders / calendar) — keep 3.x; APScheduler 4.x is still alpha
APScheduler>=3.10.0,<4.0.0

# Dev / Testing
pytest>=9.0.0
pytest-mock>=3.14.0
ruff>=0.9.0
mypy>=1.14.0
